+++ x/evm/types/params.go
@@ -41,7 +41,7 @@ func NewParams(evmDenom string) Params {
 // DefaultParams returns default evm parameters
 func DefaultParams() Params {
 	return Params{
-		EvmDenom: ethermint.DenomDefault,
+		EvmDenom: ethermint.AttoPhoton,
 	}
 }